+++ src/main/java/pv168/freelancer/ui/tablemodels/WorkDoneTableModel.java
@@ -5,6 +5,7 @@ import pv168.freelancer.model.WorkDone;
 import pv168.freelancer.model.WorkType;
 import pv168.freelancer.ui.utils.I18N;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -21,7 +22,7 @@ public class WorkDoneTableModel extends AbstractEntityTableModel<WorkDone> {
             Column.build(I18N.getString("from"), String.class, WorkDone::getFormattedWorkStart),
             Column.build(I18N.getString("to"), String.class, WorkDone::getFormattedWorkEnd),
             Column.build(I18N.getString("hours"), Long.class, WorkDone::calculateHours),
-            Column.build(I18N.getString("income"), Double.class, WorkDone::calculatePay),
+            Column.build(I18N.getString("income"), BigDecimal.class, WorkDone::calculatePay),
             Column.build(I18N.getString("note"), String.class, WorkDone::getDescription)
             );
